requirements.txt
cffi
cloudpickle==0.5.3
cycler==0.10.0
dask==0.18.2
decorator==4.3.0
imageio==2.3.0
kiwisolver
matplotlib
networkx==2.1
numpy
pandas
Pillow
pycparser==2.18
pygit==0.1
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
PyWavelets
PyYAML==5.1
scikit-image
scikit-learn
scipy
six==1.11.0
toolz==0.9.0
tqdm==4.24.0
torch
torchvision==0.2.1
